# docs/swagger.yaml
definitions:
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handlers.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handlers.TokenPair:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  model.Category:
    properties:
      created_at:
        type: string
      deleted_at:
        description: Soft delete
        type: string
      id:
        description: Adds some metadata fields to the table
        type: string
      name:
        type: string
      type:
        description: '''income'' or ''expense'''
        type: string
      updated_at:
        type: string
      userID:
        description: Foreign key to User
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/auth/callback:
    get:
      description: Handles the callback from Google OAuth, registers the user if necessary,
        and returns tokens
      responses:
        "200":
          description: Successful login
          schema:
            $ref: '#/definitions/handlers.TokenPair'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Google OAuth Callback
      tags:
      - auth
  /api/auth/email-login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with email and password, returns access and
        refresh tokens
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful login
          schema:
            $ref: '#/definitions/handlers.TokenPair'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Invalid credentials
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Login with Email and Password
      tags:
      - auth
...
  /api/auth/logout:
    get:
      description: Clears the authentication tokens, effectively logging out the user
      responses:
        "200":
          description: Successfully logged out!
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Logout
      tags:
      - auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: Uses a valid refresh token to generate a new access token
      parameters:
      - description: Refresh token request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Access token successfully refreshed
          schema:
            $ref: '#/definitions/handlers.TokenPair'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Invalid or expired refresh token
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Refresh Access Token
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Registers a new user with email and password
      parameters:
      - description: User Registration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "409":
          description: User already exists
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Register a new user
      tags:
      - auth
  /api/categories:
    get:
      consumes:
      - application/json
      description: Get categories for the authenticated user
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Category'
            type: array
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Add a new category to the database
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category to add
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/model.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Category'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to add category
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a new category
      tags:
      - categories
...
swagger: "2.0"